Change class rsolver to ANTLR inputstream rsolver
